+++ tests/test_round_trip.py
@@ -1,15 +1,13 @@
 from functools import partial
 
 import pytest
+from plumbum import ProcessExecutionError
 from plumbum.cmd import pg_dump
 
 from pg_grant import parse_acl_item, FunctionInfo, PgObjectType
 from pg_grant.query import (
     get_all_function_acls, get_all_sequence_acls, get_all_table_acls,
     get_all_type_acls)
-
-
-pytestmark = pytest.mark.nocontainer
 
 
 def _priv_acls(conn, acls, type_, revoke):
@@ -52,7 +50,14 @@ revoke_function_acls = partial(_priv_acls, type_=PgObjectType.FUNCTION, revoke=T
 def test_revoke_grant_schema_relations(connection, postgres_url, get, revoke, grant):
     cmd = pg_dump['--schema-only', postgres_url]
 
-    code, dump1, _ = cmd.run()
+    try:
+        code, dump1, err = cmd.run()
+    except ProcessExecutionError as exc:
+        if 'server version mismatch' in exc.stderr:
+            pytest.skip("pg_dump is older than server version")
+        else:
+            raise
+
     assert code == 0
 
     acls = get(connection, 'public')