+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Change Log
 
 ## [Unreleased][unreleased]
-N/A
+### Fixed
+- ACLs for functions without any arguments (i.e. `arg_types=()`) could not be
+  queried using `get_function_acl`. 
 
 ## [0.3.2][] - 2020-05-21
 ### Fixed